Fix the CI test coverage failure.
